refactor: unify entrypoint and logging
