--- __main__.py
@@ -1,108 +1,0 @@
-"""Module entrypoint for ``python -m ai-trading-bot``."""
-
-from __future__ import annotations
-
-import logging
-import warnings
-import os
-import time
-from filelock import FileLock, Timeout
-
-# AI-AGENT-REF: suppress noisy external library warnings
-warnings.filterwarnings("ignore", category=SyntaxWarning, message="invalid escape sequence")
-warnings.filterwarnings("ignore", message=".*_register_pytree_node.*")
-
-try:
-    from alpaca_trade_api.rest import REST
-except Exception:  # pragma: no cover - optional dependency
-    REST = object  # type: ignore
-
-from run import main as entrypoint
-import config
-from logger import get_logger
-from data_fetcher import DataFetchError as DataSourceEmpty
-
-# AI-AGENT-REF: throttle HEALTH_ROWS and SKIP_COOLDOWN logs
-_LAST_HEALTH_LOG_TIME = 0.0
-_skip_cooldown_symbols: set[str] = set()
-_LAST_SKIP_LOG_TIME = 0.0
-_last_skip_set: set[str] = set()
-
-
-def check_health_rows(rows: int) -> None:
-    """Log HEALTH_ROWS at most every 10 seconds."""
-    global _LAST_HEALTH_LOG_TIME
-    now = time.monotonic()
-    if now - _LAST_HEALTH_LOG_TIME > 10:
-        get_logger(__name__).debug("HEALTH_ROWS", extra={"rows": rows})
-        _LAST_HEALTH_LOG_TIME = now
-
-
-def skip_cooldown(symbols: list[str]) -> None:
-    # Enforce immediate cooldown protection to avoid double buying
-    for sym in symbols:
-        if sym in positions:
-            continue  # Already holding, skip re-buy
-        cooldown_tracker[sym] = time.time()
-    """Batch SKIP_COOLDOWN symbols until flushed."""
-    global _skip_cooldown_symbols
-    _skip_cooldown_symbols.update(symbols)
-
-
-def flush_skip_cooldown_log() -> None:
-    global _skip_cooldown_symbols, _LAST_SKIP_LOG_TIME, _last_skip_set
-    if not _skip_cooldown_symbols:
-        return
-    now = time.monotonic()
-    if (
-        _skip_cooldown_symbols != _last_skip_set
-        or now - _LAST_SKIP_LOG_TIME >= 15
-    ):
-        get_logger(__name__).info(
-            "SKIP_COOLDOWN_BATCHED | %s",
-            ", ".join(sorted(_skip_cooldown_symbols)),
-        )
-        _LAST_SKIP_LOG_TIME = now
-        _last_skip_set = set(_skip_cooldown_symbols)
-    _skip_cooldown_symbols.clear()
-
-
-def main() -> None:  # pragma: no cover - thin wrapper for entrypoint
-    """Invoke :func:`run.main` when module is executed as a package."""
-    os.environ.setdefault("BACKTEST_SERIAL", "1")  # AI-AGENT-REF: ensure serial backtests
-    if config.FORCE_TRADES:
-        get_logger(__name__).warning(
-            "\ud83d\ude80 FORCE_TRADES is ENABLED. This run will ignore normal health halts!"
-        )
-    lock = FileLock("/tmp/ai_trading_bot.lock", timeout=0)
-    try:
-        with lock:
-            for attempt in range(1, 4):
-                try:
-                    entrypoint()
-                    break
-                except DataSourceEmpty:
-                    get_logger(__name__).warning(
-                        "DATA_SOURCE_EMPTY, retrying %d/3…",
-                        attempt,
-                    )
-                    time.sleep(attempt * 2)
-            else:
-                get_logger(__name__).error(
-                    "DATA_SOURCE_EMPTY after 3 attempts; skipping trade cycle"
-                )
-                return
-    except Timeout:
-        get_logger(__name__).info("RUN_ALL_TRADES_SKIPPED_OVERLAP")
-        return
-
-
-if __name__ == "__main__":
-    logging.basicConfig(
-        format="%(asctime)sZ %(levelname)s [%(bot_phase)s] %(name)s - %(message)s",
-        datefmt="%Y-%m-%dT%H:%M:%S",
-        level=logging.INFO,
-    )
-    logging.Formatter.converter = lambda *args: time.gmtime()  # force UTC
-    main()
-
