+++ tests/test_additional_coverage.py
@@ -16,7 +16,7 @@ import config
 import meta_learning
 import ml_model
 import risk_engine
-import run as main
+import main
 import utils
 import validate_env
 from strategies.mean_reversion import MeanReversionStrategy
@@ -79,7 +79,7 @@ def test_create_flask_routes():
 
 def test_main_serve_api(monkeypatch):
     """main launches Flask thread and bot when --serve-api used."""
-    monkeypatch.setattr(sys, "argv", ["run.py", "--serve-api"])
+    monkeypatch.setattr(sys, "argv", ["main.py", "--serve-api"])
     monkeypatch.setattr(main, "run_bot", lambda v, s: 0)
     called = {}
